// fe/pwandora/monitor/src/types/confirmationRefs.ts
export type IConfirmationRefs = {
  crawledData: React.RefObject<HTMLDivElement | null>
  aiSuggestions: React.RefObject<HTMLDivElement | null>
  screenshots: React.RefObject<HTMLDivElement | null>
  displayCompatibility: React.RefObject<HTMLDivElement | null>
  aiCensor: React.RefObject<HTMLDivElement | null>
  packaging: React.RefObject<HTMLDivElement | null>
  lighthouseTest: React.RefObject<HTMLDivElement | null>
  security: React.RefObject<HTMLDivElement | null>
}